Add CockroachDB integration tests and use CockroachDB in examples
This commit updates the ent codebase to demonstrate support for CockroachDB. It
adds new integration tests for CockroachDB that live alongside the existing
Postgres integration tests. It also replaces the use of SQLite in the example
applications (`./examples`) with CockroachDB.

There are a few interesting things in here that I was able to tweak to get this
working:
- certain schema changes get tripped up without `default_int_size=4`
- certain joins never complete planning without `reorder_joins_limit=4`
- CockroachDB does not support the macaddr data type
- FOR SHARE is a no-op in CockroachDB
- transactional schema changes block in some cases in CockroachDB. I did not
  bottom out on this.

However, the biggest blocker for all of this is support for `GENERATED BY
DEFAULT AS IDENTITY`. Luckily, this is close to supported in CockroachDB due to
the fantastic work in https://github.com/cockroachdb/cockroach/pull/69107. With
that patched in, everything here runs.

To run the tests or the examples, simply spin up a demo cluster alongside with
`cockroach demo --insecure --empty`.